[Navigation API] Split NavigationAPIMethodTracker, the method tracker registry, the rate limiter, and the promise settlement observer into their own files
Reviewed by NOBODY (OOPS!).

Navigation.cpp had grown to hold several self-contained helpers inline. This moves the
components that are already conceptually separate out of Navigation.{h,cpp} to reduce the
file size and make the remaining Navigation logic easier to follow. No behavior change.

- NavigationAPIMethodTracker and its registry (formerly the nested
  Navigation::MethodTrackerRegistry, now the top-level NavigationMethodTrackerRegistry)
  move to page/NavigationAPIMethodTracker.{h,cpp}.
- The rate limiter (formerly Navigation::RateLimiter) moves to page/NavigationRateLimiter.h,
  with navigationAllowed() inlined so the class stays header-only.
- PromiseSettlementObserver, a generic WebIDL wait-for-all primitive that is not
  Navigation-specific, moves to dom/PromiseSettlementObserver.{h,cpp}.

Covered by existing tests (imported/w3c/web-platform-tests/navigation-api,
LayoutTests/navigation-api, http/wpt/navigation-api, http/tests/navigation-api).

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/dom/PromiseSettlementObserver.h: Added.
* Source/WebCore/dom/PromiseSettlementObserver.cpp: Added.
(WebCore::PromiseSettlementObserver::create):
(WebCore::PromiseSettlementObserver::PromiseSettlementObserver):
(WebCore::PromiseSettlementObserver::waitForAll):
(WebCore::PromiseSettlementObserver::resolve):
(WebCore::PromiseSettlementObserver::reject):
(WebCore::PromiseSettlementObserver::handleResult):
(WebCore::PromiseSettlementObserver::registerPromise):
* Source/WebCore/page/NavigationAPIMethodTracker.h: Added.
* Source/WebCore/page/NavigationAPIMethodTracker.cpp: Added.
(WebCore::NavigationAPIMethodTracker::create):
(WebCore::NavigationAPIMethodTracker::NavigationAPIMethodTracker):
(WebCore::NavigationAPIMethodTracker::commitTo):
(WebCore::NavigationAPIMethodTracker::resolveFinished):
(WebCore::NavigationAPIMethodTracker::rejectFinished):
(WebCore::NavigationMethodTrackerRegistry::setUpcomingNonTraverse):
(WebCore::NavigationMethodTrackerRegistry::addUpcomingTraverse):
(WebCore::NavigationMethodTrackerRegistry::upcomingTraverse):
(WebCore::NavigationMethodTrackerRegistry::ongoing):
(WebCore::NavigationMethodTrackerRegistry::takeUpcomingNonTraverseIfEquals):
(WebCore::NavigationMethodTrackerRegistry::promoteUpcomingNonTraverseToOngoing):
(WebCore::NavigationMethodTrackerRegistry::promoteUpcomingTraverseToOngoing):
(WebCore::NavigationMethodTrackerRegistry::unregister):
(WebCore::NavigationMethodTrackerRegistry::isEmpty):
(WebCore::NavigationMethodTrackerRegistry::visitInGCThread):
* Source/WebCore/page/NavigationRateLimiter.h: Added.
(WebCore::NavigationRateLimiter::navigationAllowed):
* Source/WebCore/page/Navigation.h:
* Source/WebCore/page/Navigation.cpp:
