+++ Source/WebCore/Sources.txt
@@ -1400,6 +1400,7 @@ dom/PositionIterator.cpp @header:RenderStyleGetters
 dom/ProcessingInstruction.cpp
 dom/ProgressEvent.cpp
 dom/PromiseRejectionEvent.cpp
+dom/PromiseSettlementObserver.cpp
 dom/PseudoElement.cpp @header:RenderStyleGetters
 dom/QualifiedName.cpp
 dom/QualifiedNameCache.cpp
@@ -2242,6 +2243,7 @@ page/MemoryRelease.cpp @header:RenderStyleGetters @cost:9
 page/MouseEventWithHitTestResults.cpp
 page/NavigateEvent.cpp
 page/Navigation.cpp
+page/NavigationAPIMethodTracker.cpp
 page/NavigationActivation.cpp
 page/NavigationCurrentEntryChangeEvent.cpp
 page/NavigationDestination.cpp
