+++ Source/WebCore/dom/PromiseSettlementObserver.h
@@ -1,0 +1,68 @@
+/*
+ * Copyright (C) 2024-2025 Apple Inc. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS''
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+ * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS
+ * BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
+ * THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#pragma once
+
+#include <JavaScriptCore/JSCJSValue.h>
+#include <wtf/Forward.h>
+#include <wtf/Function.h>
+#include <wtf/Ref.h>
+#include <wtf/RefCounted.h>
+#include <wtf/Vector.h>
+
+namespace WebCore {
+
+class Document;
+class DOMPromise;
+
+// https://webidl.spec.whatwg.org/#wait-for-all
+class PromiseSettlementObserver : public RefCounted<PromiseSettlementObserver> {
+public:
+    using SuccessSteps = Function<void()>;
+    using FailureSteps = Function<void(JSC::JSValue)>;
+
+    static Ref<PromiseSettlementObserver> create(SuccessSteps&&, FailureSteps&&);
+
+    // https://webidl.spec.whatwg.org/#wait-for-all
+    void waitForAll(Document&, const Vector<Ref<DOMPromise>>&);
+
+private:
+    PromiseSettlementObserver(SuccessSteps&&, FailureSteps&&);
+
+    bool isWaiting() const { return m_totalPromises > m_settledPromises; }
+
+    void resolve();
+    void reject(JSC::JSValue result);
+    void handleResult(bool isFulfilled, JSC::JSValue result);
+    bool registerPromise(DOMPromise&);
+
+    SuccessSteps m_successSteps;
+    FailureSteps m_failureSteps;
+    bool m_settled { false };
+    unsigned m_totalPromises { 0 };
+    unsigned m_settledPromises { 0 };
+};
+
+} // namespace WebCore
